Amistades ya funciona correctamente

+++ Chally/app/Http/Controllers/AmistadController.php
@@ -6,38 +6,40 @@ use Auth;
 use App\Usuario;
 use Illuminate\Http\Request;
 use App\Amistad;
-use Carbon\Carbon;  
+use Carbon\Carbon;
+use Illuminate\Support\Facades\Auth as FacadesAuth;
 
 class AmistadController extends Controller
 {
     /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    * Display a listing of the resource.
+    *
+    * @return \Illuminate\Http\Response
+    */
     public function index()
     {
         //
     }
-
+    
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    * Show the form for creating a new resource.
+    *
+    * @return \Illuminate\Http\Response
+    */
     public function create()
     {
         //
     }
-
+    
     /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+    * Store a newly created resource in storage.
+    *
+    * @param  \Illuminate\Http\Request  $request
+    * @return \Illuminate\Http\Response
+    */
     public function store($username)
     {
+        //Creamos en la db una nueva fila que va a indicar que se mando la solicitud de amistad
         $id_usuario_1 = Auth::user()->id_usuario;
         $usuario_2 = Usuario::where('username', $username)->first();
         $id_usuario_2 = $usuario_2->id_usuario;
@@ -46,56 +48,58 @@ class AmistadController extends Controller
         $nuevaAmistad->id_usuario_2 = $id_usuario_2;
         $nuevaAmistad->created_at = Carbon::now();
         $nuevaAmistad->updated_at = null;
-
+        
         $nuevaAmistad->save();
-
+        
         return redirect("/usuario/$username");
     }
-
+    
     static function verificarAmistad($id_usuario_1, $id_usuario_2)
     {
-        $amistad = Amistad::where('id_usuario_1', $id_usuario_1)->first();
+        if($id_usuario_2 == $id_usuario_1){
+            return "persona";
+        }
 
+        $amistad = Amistad::where('id_usuario_1', $id_usuario_1)->where('id_usuario_2' , $id_usuario_2)->first();
+        
         if ($amistad) {
+            
             if ($amistad->id_usuario_2 == $id_usuario_2 && $amistad->updated_at == null) {
                 //solicitud enviada pero no aceptada
                 return "enviada";
             }
-
+            
             if ($amistad->id_usuario_2 == $id_usuario_2 && $amistad->updated_at != null) {
                 return "amigos";
             }
-
-            if($id_usuario_2 == $id_usuario_1){
-                return "persona";
-            }
+        }else{
+            return "not amigos";
         }
-
-        return "not amigos";
+        
     }
-
+    
     /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    * Display the specified resource.
+    *
+    * @param  int  $id
+    * @return \Illuminate\Http\Response
+    */
     public function show($id)
     {
         //
     }
-
+    
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    * Show the form for editing the specified resource.
+    *
+    * @param  int  $id
+    * @return \Illuminate\Http\Response
+    */
     public function edit($username)
     {
         $usuario = Usuario::where('username' , $username)->first();
         $id_usuario_2 = $usuario->id_usuario;
-
+        
         //Todas las solicitudes son aquellas que tengan updated_at en null 
         $solicitudes = Amistad::all()->where('id_usuario_2' , $id_usuario_2)->where('updated_at' , NULL);
         $usuarios = [];
@@ -109,19 +113,19 @@ class AmistadController extends Controller
         }
         
     }
-
+    
     /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update($username , $estado)
+    * Update the specified resource in storage.
+    *
+    * @param  \Illuminate\Http\Request  $request
+    * @param  int  $id
+    * @return \Illuminate\Http\Response
+    */
+    public function update($estado , $username )
     {
         $usuario = Usuario::where('username' , $username)->first();
         $id_usuario_1 = $usuario->id_usuario;
-        $amistad = Amistad::where('id_usuario_1' , $id_usuario_1)->where('id_usuario_2' , Auth::user()->id_usuario)->first();
+        $amistad = Amistad::where('id_usuario_1' ,  Auth::user()->id_usuario)->where('id_usuario_2' , $id_usuario_1)->first();
         $id_amistad = $amistad->id_amistad;
         if($estado == 'aceptar'){
             $amistad->updated_at = Carbon::now();
@@ -131,23 +135,24 @@ class AmistadController extends Controller
             $amistad_2->created_at = Carbon::now();
             $amistad_2->updated_at = Carbon::now();
             $amistad_2->save();
-        }else{
+        }elseif($estado == 'cancelar'){
             $amistad->destroy($id_amistad);
+            return redirect("/usuario/" . $username);
         }
-
-            $amistad->save();
-            
-
-
+        
+        $amistad->save();
+        
+        
+        
         return redirect("/usuario/" . Auth::user()->username . "/solicitudes");
     }
-
+    
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    * Remove the specified resource from storage.
+    *
+    * @param  int  $id
+    * @return \Illuminate\Http\Response
+    */
     public function destroy($id)
     {
         //
